Move core types inline
