+++ lib/react/react.tsx
@@ -1,7 +1,6 @@
 import type { ReactNode } from 'react'
 import React, { createContext, useContext } from 'react'
-import type { FeathersClient } from '../adapters/feathers.js'
-import { Figbird } from '../core/figbird.js'
+import type { Figbird } from '../core/figbird.js'
 import type { AnySchema, Schema } from '../schema/types.js'
 
 // Create a generic context factory to maintain type safety
@@ -18,20 +17,6 @@ export function useFigbird<S extends Schema = AnySchema>(): Figbird<S> {
     throw new Error('useFigbird must be used within a FigbirdProvider')
   }
   return context
-}
-
-/**
- * Specific to Feathers adapter. Might remove in the future.
- */
-export function useFeathers(): FeathersClient {
-  const figbird = useFigbird()
-  const adapter = figbird.adapter as { feathers?: FeathersClient }
-
-  if (!adapter?.feathers) {
-    throw new Error('useFeathers must be used with a Feathers adapter')
-  }
-
-  return adapter.feathers
 }
 
 interface FigbirdProviderProps<S extends Schema = AnySchema> {
